% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mirt_objects.R
\name{create_mirt_MZIGRM}
\alias{create_mirt_MZIGRM}
\title{Create A Multidimensional Zero-Inflated Graded Response Model mirt custom item}
\usage{
create_mirt_MZIGRM(K)
}
\arguments{
\item{K}{Number of Item Response Categories (integer)}
}
\value{
A mirt custom item from mirt::createItem() named "MZIGRM_K"
}
\description{
IN TESTING: Pk seems correct but mirt estimation is very off
}
