% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MZIGRM_tools.R
\name{expect_prop_Z}
\alias{expect_prop_Z}
\title{Calculates expected proportion of zeros for MZI-GRM}
\usage{
expect_prop_Z(slope, intercept, theta = c(0, 0), D = 1.702)
}
\arguments{
\item{slope}{A J x 2 matrix of item slopes}

\item{intercept}{A J x K-1 matrix of item intercpets}

\item{theta}{A vector of susceptibility and severity latent scores; defaults to the mean 0, 0}

\item{D}{A scaling constant for the GRM}
}
\value{
A J x K matrix
}
\description{
Calculates the expected proportion of zero responses for a respondent of a given theta level given item parameters.
Also provides the proportion of those zeros attributable to 1) susceptibility and 2) severity
}
