% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHGRM_tools.R
\name{generate_MHGRM_data}
\alias{generate_MHGRM_data}
\title{Generate MHGRM Data}
\usage{
generate_MHGRM_data(
  N,
  J,
  K,
  rho = 0,
  theta = NULL,
  a = NULL,
  b = NULL,
  seed_person_params = NULL,
  seed_item_params = NULL,
  seed_response = NULL
)
}
\arguments{
\item{N}{Number of Observations}

\item{J}{Number of Items}

\item{K}{Number of Response Categories. Must be greater than 2 for the MHGRM}

\item{rho}{Correlation Between Person Parameters}

\item{theta}{An Nx2 matrix of Person Parameters}

\item{a}{A Jx2 matrix of slope parameters}

\item{b}{A JxK matrix of intercept parameters}

\item{seed_person_params}{An optional numeric to set the seed for person parameters}

\item{seed_item_params}{An optional numeric to set the seed for item parameters}

\item{seed_response}{An optional numeric or J length numeric to set the seed for item responses}
}
\value{
A list containing the Generated Person Parameters, Item Parameters, and simulate item responses
}
\description{
Generates Some Factory MHGRM Item Parameters, Person Parameters, and Item Responses
}
