% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHGRM_tools.R
\name{GRM_mat2}
\alias{GRM_mat2}
\title{GRM Item Response Probabilities}
\usage{
GRM_mat2(theta, slope, intercept, D = 1.702)
}
\arguments{
\item{theta}{An N x M matrix of person parameters}

\item{slope}{A 1 x M matrix of item slopes}

\item{intercept}{A 1 x K-1 matrix of item intercepts}

\item{D}{A scaling Constant}
}
\value{
An N x K matrix of item response probabilities
}
\description{
Calculates the Graded Response Model response category probabilities
}
