% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHGRM_tools.R
\name{MHGRM_mat2}
\alias{MHGRM_mat2}
\title{MHGRM Response Probabilities}
\usage{
MHGRM_mat2(theta, slope, intercept, D = 1.702)
}
\arguments{
\item{theta}{An N x 2 Matrix of Person Parameters}

\item{slope}{A 1 x 2 Matrix of item slopes}

\item{intercept}{A 1 x K-1 Matrix of Item Intercepts}

\item{D}{A scaling Constant}
}
\value{
An N x K matrix of response probabilities
}
\description{
Calculates the MH Graded Response Model response category probabilities
}
