% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHGRM_tools.R
\name{MHGRM_simulate_response}
\alias{MHGRM_simulate_response}
\title{Simulate Item Responses from MHGRM}
\usage{
MHGRM_simulate_response(x, seed = NULL)
}
\arguments{
\item{x}{An N x K matrix of response probabilities}

\item{seed}{A numeric value to set the random seed}
}
\value{
An N length vector of item responses ranging from 0:(K-1)
}
\description{
Simulates item responses under the MHGRM give an N x K matrix of response probabilities
}
