% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{MZI_dat}
\alias{MZI_dat}
\title{Multidimensional Zero Inflated GRM Data}
\format{
A data frame with 500 rows and 12 variables
\describe{
\item{item1,item2,item3,item4,item5,item6,item7,item8,item9,item10}{item responses 0 - 4}
\item{Presence}{latent presence score}
\item{Severity}{latent severity score}
}
}
\usage{
MZI_dat
}
\description{
A simulated data set with 500 respondents and 10 items for the MZI-GRM
}
\keyword{datasets}
