% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MZIGRM_tools.R
\name{MZI_propZI}
\alias{MZI_propZI}
\title{Calculates Expected Proportion of Zero Inflation}
\usage{
MZI_propZI(b0, b1)
}
\arguments{
\item{b0}{The item intercept associated the theta0 (presence)}

\item{b1}{The item intercepts associated with a zero response from theta1 (severity)}
}
\value{
A numeric
}
\description{
Calculates the expected proportion of zeros due to zero inflation as a function of the presence latent variable. Assumes theta is standard bivariate normal.
}
