% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MZIGRM_tools.R
\name{MZIGRM_mat2}
\alias{MZIGRM_mat2}
\title{MZIGRM Response Probabilities}
\usage{
MZIGRM_mat2(theta, slope, intercept, D = 1.702)
}
\arguments{
\item{theta}{An N x 2 Matrix of Person Parameters}

\item{slope}{A 1 x 2 Matrix of item slopes}

\item{intercept}{A 1 x K Matrix of Item Intercepts. The first intercept corresponds to the 2PL component}

\item{D}{A scaling Constant}
}
\value{
An N x K matrix of response probabilities
}
\description{
Calculates the MZI Graded Response Model response category probabilities
}
