% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MZIGRM_tools.R
\name{MZIGRM_simulate_response}
\alias{MZIGRM_simulate_response}
\title{Simulate Item Responses for the MZIGRM}
\usage{
MZIGRM_simulate_response(x, seed = NULL)
}
\arguments{
\item{x}{An N x K matrix of response probabilities}

\item{seed}{A numeric value to set the random seed}
}
\value{
An N length vector of item responses 0:(K-1)
}
\description{
Simulate Item Responses for the MZIGRM
}
