% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{STB}
\alias{STB}
\title{Suicidal Thoughts and Behaviors Data}
\format{
A data frame with 1744 rows and 23 variables
\describe{
\item{ID}{An ID Variable}
\item{Gender}{Self-reported Gender}
\item{Age}{Self-reported age in years}
\item{Dx}{Self-reported history of any mental health diagnosis}
\item{STBhx}{Self-reported history of any suicidal thoughts or behaviors}
\item{PASIS1,PASIS2,PASIS3,PASIS4,PASIS5,PASIS6,PASIS7,PASIS8,PASIS9,
PASIS10,PASIS11,PASIS12,PASIS13,PASIS14,PASIS15,PASIS16,PASIS17,PASIS18}{Responses to the Passive and Active Suicidal Ideation Scale Modified for Past Day Reference}
}
}
\usage{
STB
}
\description{
Data on past day suicidal ideation from McClure et al., 2025
}
\keyword{datasets}
